Refactor: Separate Carousel Component
